refactor: canvasへのレンダー処理を切り出した
